Fix the stack buttons, paste files and images, and collapse twins
- give the stack footer real symbols: the push icon named one that
  does not exist and rendered as an empty pill
- take a pasted file or screenshot the way a drop is taken, staged
  and typed into the agent
- disclose for a stack entry with no recorded live session from its
  newest conversation, on the pickers' defaults
- keep entries above an unpushed branch out of reach in the pull
  request tab, since nothing can be listed or opened for them yet;
  the review strip still shows them
- treat two branches at one commit as one entry: the checked-out
  name, else the one the remote knows, else the one created first,
  so a rename's leftover or a mistaken copy never becomes a stack
