+++ App/RootView+UtilityTabs.swift
@@ -11,11 +11,16 @@ extension RootView {
     /// The agent terminal: copies are prose, so multi-line copies
     /// reflow for pasting into chat and pull request bodies.
     func agentTerminal(for session: AgentSession, at worktreePath: String, isActive: Bool) -> some View {
-        TerminalPaneView(
+        // A pasted file or screenshot reaches the agent the way a
+        // dropped one does. Bound first: as the call's last argument
+        // the formatter would make it trailing, which fights SwiftLint.
+        let pasteFiles: ([URL]) -> Bool = { urls in dropFiles(urls, into: session.name) }
+        return TerminalPaneView(
             command: session.paneID.map(dependencies.service.attachCommand(paneID:)) ?? [],
             reflowsCopies: true,
             isActive: isActive,
             fixedAppearance: dependencies.service.launchAppearance(worktreePath: worktreePath),
+            onPasteFiles: pasteFiles,
         )
         // A hair of room either side: the agent's own frames draw to
         // their last column, which sat against the pane's edges.
